requirements.txt
mss
Pillow
opencv-python
pyzbar
numpy
imagehash
requests
pyobjc-framework-Quartz
